+++ .gitignore
@@ -186,7 +186,7 @@ tests/acls/acl.sh
 tests/urls/request.sh
 tests/alarm_repetition/alarm.sh
 tests/template_dimension/template_dim.sh
-aclk/tests/install-fake-charts.d.sh
+aclk/legacy/tests/install-fake-charts.d.sh
 
 # tests and temp files
 test-driver
