+++ Makefile.am
@@ -116,7 +116,7 @@ SUBDIRS += \
     web \
     claim \
     parser \
-    aclk \
+    aclk/legacy \
     spawn \
     $(NULL)
 
@@ -524,27 +524,27 @@ PARSER_FILES = \
     $(NULL)
 
 ACLK_FILES = \
-    aclk/aclk_rrdhost_state.h \
-    aclk/aclk_common.c \
-    aclk/aclk_common.h \
-    aclk/aclk_stats.c \
-    aclk/aclk_stats.h
+    aclk/legacy/aclk_rrdhost_state.h \
+    aclk/legacy/aclk_common.c \
+    aclk/legacy/aclk_common.h \
+    aclk/legacy/aclk_stats.c \
+    aclk/legacy/aclk_stats.h \
     $(NULL)
 
 if ENABLE_ACLK
 ACLK_FILES += \
-    aclk/agent_cloud_link.c \
-    aclk/agent_cloud_link.h \
-    aclk/aclk_query.c \
-    aclk/aclk_query.h \
-    aclk/mqtt.c \
-    aclk/mqtt.h \
-    aclk/aclk_rx_msgs.c \
-    aclk/aclk_rx_msgs.h \
-    aclk/aclk_lws_wss_client.c \
-    aclk/aclk_lws_wss_client.h \
-    aclk/aclk_lws_https_client.c \
-    aclk/aclk_lws_https_client.h \
+    aclk/legacy/agent_cloud_link.c \
+    aclk/legacy/agent_cloud_link.h \
+    aclk/legacy/aclk_query.c \
+    aclk/legacy/aclk_query.h \
+    aclk/legacy/mqtt.c \
+    aclk/legacy/mqtt.h \
+    aclk/legacy/aclk_rx_msgs.c \
+    aclk/legacy/aclk_rx_msgs.h \
+    aclk/legacy/aclk_lws_wss_client.c \
+    aclk/legacy/aclk_lws_wss_client.h \
+    aclk/legacy/aclk_lws_https_client.c \
+    aclk/legacy/aclk_lws_https_client.h \
     $(NULL)
 endif
