+++ aclk/legacy/aclk_rx_msgs.h
@@ -3,7 +3,7 @@
 #ifndef NETDATA_ACLK_RX_MSGS_H
 #define NETDATA_ACLK_RX_MSGS_H
 
-#include "../daemon/common.h"
+#include "../../daemon/common.h"
 #include "libnetdata/libnetdata.h"
 
 int aclk_handle_cloud_message(char *payload);
